+++ shipyard_blueprints/notion/shipyard_notion/__init__.py
@@ -1,4 +1,4 @@
-from .notion_db import NotionClient
+from .notion import NotionClient
 from .notion_utils import (
     convert_pandas_to_notion,
     mapper,
